+++ space/Tesselator.cpp
@@ -39,7 +39,7 @@ bool Tesselator::test(OctreeNodeData &params) {
 	return !params.node->isSolid() && !params.node->isSimplified();
 }
 
-void Tesselator::getOrder(OctreeNodeData &params, int * order){
+void Tesselator::getOrder(OctreeNodeData &params, uint8_t * order){
 	for(int i = 7 ; i >= 0 ; --i) {
 		order[i] = i;
 	}
